CFY-7656 Always convert a CA certificate to pkcs12
Whether the CA cert was provided by the user or generated, it must
always be also copied to a PKCS12 form.
That PKCS12 bundle will only be used for verifying requests, and it
must not contain the private key (which might not even be available
in case of a user-provided CA cert).