Sign external certificates with the CA cert/key (#104)
* Sign external certificates with the CA cert/key

This should allow the users to pass a single CA cert/key pair as the input for the installation and distribute the CA cert to the endpoints (UI/CLI). This should make working in HA clusters much simpler for the users.

* Use the CA cert in the CLI

When generating the external cert we're using the CA cert to sign it, so it would be the only certificate that could be used for verification.