CFY-7656 Generate PKCS12 from internal cert + key
Instead of passing the CA cert + key (which might not be present)
into the PKCS12 bundle (for riemann to use), provide the internal
cert + key, which is enough for clojure's http client to validate
the connection.

Note that this is unlike openssl-based clients, which would need
the CA cert, not the server cert itself.