CY-694 Only use generated security configuration if --clean-db is set  (#357)
* CY-694 Factor out a method to chown the resources dir

* CY-694 Only use generated security configuration if --clean-db is set
Because the encryption key and other values in the security
configuration are tightly coupled to values in the DB (e.g. rabbitmq
passwords, user passwords), we shouldn't overwrite rest-security.conf
every time we run configure/install. If `--clean-db` is set, then the
DB is cleansed anyway, and in this case we'll use the generated config.
If not, we'll merge the dict in config.yaml with the JSON file already
on the manager, because we want the old values, but also want any new
keys to exist in the JSON.