Move comment to avoid ruamel bug
With the comment in its former position, it was sometimes being duplicated
and the following two lines were being rolled into one, which made the
YAML invalid. This seems to reliably happen with some combination of config
settings, but not with others, and only when run through cfy_manager install.