mgmtworker: move starting into .start()
The main thing here is removing the cluster.configure() call.
This is because that call does absolutely nothing other than
verifications; the actual joining to a cluster happens elsewhere
(restservice).

The cluster.py file should be reworked or removed, but I'm going
to say this is outside the scope of this.

There is CY-2796 about fixing up the remove-from-cluster flow,
which is going to be about reworking cluster.py indeed, and making
remove there actually work.