Separate configure & start calls (#776)
* Call components .start() while installing

Currently, when installing (ie. cfy_manager install,
without --only-install), each component is installed, and then
each component's .configure() method is called.

And each component's .configure() method happens to also start it.

That is just bonkers. Instead, after calling .configure() for all
components, also call .start() for all components.

Then, we'll need to make sure that each configure method only
configures it, and doesn't start it.

* rabbitmq: move starting into .start()

...as opposed to starting in .configure()

Note that .join_cluster() currently happens in start, because
well, it needs to happen after rabbitmq has been started.
But really, it feels like its own separate step.

* restservice: move starting into .start()

Things like uploading the license are obvious that they must
happen after starting.

The big change here is moving ._configure_db(), because that is
where all the create_db_tables_and_add_default.py stuff happens,
ie. where the tables are populated, amqp users are created, etc.
That means this must run after the db has been started, and after
rabbitmq has been started.

All the cluster stuff still feels like it should be somehow separate.

* stage: move starting into .start()

Main point here is that stage does its db migrations here, so
that must happen after the db has been started.

* composer: move starting into .start()

same as stage

* mgmtworker: move starting into .start()

The main thing here is removing the cluster.configure() call.
This is because that call does absolutely nothing other than
verifications; the actual joining to a cluster happens elsewhere
(restservice).

The cluster.py file should be reworked or removed, but I'm going
to say this is outside the scope of this.

There is CY-2796 about fixing up the remove-from-cluster flow,
which is going to be about reworking cluster.py indeed, and making
remove there actually work.

* cli: rename .configure() to .start()

It's not really "configure", and it's not really "start".
Anyway, it requires restservice to be started, since that's
what `cfy profiles use` does.

* sanity: rename .configure() to .start()

It's not really "configure" and it's not really "start".
Sanity-check doesn't really fit well into the 'component' model.
Anyway, it needs to run after everything's been started.

* nginx: move cert handling into .start()

This is the part that is unfortunate. Anyway, we need this to be
inside of start, so that it only happens after a (possible)
cluster join.

The CA cert is generated in rest's start, because the cluster
code path does something different.
Hopefully someday we'll be able to clean this up. But I don't want
to do this level of refactors inside of this pr.

* Remove cluster.py

* Inline a method in composer and stage