Make a new dockerfile for building the manager image
The point is having a Dockerfile that comes with the manager
preinstalled; when started, the container will run
`cfy_manager configure` and `cfy_manager start`.

The point is that the configure&start will then take into account
the config.yaml which was perhaps mounted by the user.
This allows actually setting configuration for your manager docker
containers.

The way this is done, a simple oneshot service is added, which
runs those cfy_manager commands (the cloudify-starter).