Move configuring the status reporter out of restservice
The only reason it was in restservice, was to run the config after
the cert has been deployed.

The easier thing to do is to keep stat-rep's configuration inside
of the stat-rep component, and just move the component.

Also note: moving things to after restservice+nginx+rabbitmq
is good, because once those three have started, the manager is ready
to serve requests. So this makes the manager be soft-started
(ie. able to work, but not everything has been started yet) faster.