In an AIO image, install all services ever.
This allows us to only ship one image, and use it in any role:
AIO, manager, db, or queue. The image will be a bit bigger, because
it includes eg. patroni, but that's a good price to pay, since
those cluster-only packages are negligible in size compared to the
image total