CY-2755 Make a new dockerfile for building the manager image (#780)
* Make a new dockerfile for building the manager image

The point is having a Dockerfile that comes with the manager
preinstalled; when started, the container will run
`cfy_manager configure` and `cfy_manager start`.

The point is that the configure&start will then take into account
the config.yaml which was perhaps mounted by the user.
This allows actually setting configuration for your manager docker
containers.

The way this is done, a simple oneshot service is added, which
runs those cfy_manager commands (the cloudify-starter).

* Separate validating dependencies out of the main validate method

No need to run that on configure. Not running it saves on the
order of 5 seconds.

To be honest, no need to run it in install either, because we
have the rpm dependencies declared. But I'm not going to remove it
outright just yet.

* Move configuring the status reporter out of restservice

The only reason it was in restservice, was to run the config after
the cert has been deployed.

The easier thing to do is to keep stat-rep's configuration inside
of the stat-rep component, and just move the component.

Also note: moving things to after restservice+nginx+rabbitmq
is good, because once those three have started, the manager is ready
to serve requests. So this makes the manager be soft-started
(ie. able to work, but not everything has been started yet) faster.

* Implement a way to not have to overwrite the config

With this, config won't be overwritten. This has to do with the build,
we don't need to write the config file there.

For now, I'll leave this undocumented, and instead, let's just
make this the default (and only) behaviour.

* Run the status reporter as the cfyuser group

This is probably a bad idea. Maybe we can move the ssl cert to
a location that is owned by cfyreporter instead?

BTW, no idea how does this work on master

* cfy_manager start: also set settings

Settings (private ip, the "globals", etc) are actually used in start,
so we need to set them.
This was broken but we didn't notice, because we never test running
`cfy_manager configure` and then a separate `cfy_manager start`

* fixup validate

* Restservice: also set settings in start

Similar to the previous commit, things that we set at runtime in
configure, also need to be set at runtime in start, just in case
we run configure/start separately

* Rabbitmq: also set settings in start

Same as the previous commit, config entries that are set at runtime
in configure, must also be set in start.

BTW, remember that I wanted to split the config into a "config" part
and the "context" (ie. runtime) part? Really wish that was already
done.

* Rabbitmq: don't restart after setting policies

????? there's no reason to do that
We don't do it in the cluster path, either.
Only thing this does is cost us 10 seconds.

* Restservice: drop the request validation

This validation wasn't _bad_, it was just not worth it.

The cost is that, just after the restservice has been started,
gunicorn needs quite some time to actually import everything, start
up, spawn subprocesses, and respond to the request. This validation
therefore cost us on the order of ~8 seconds.

Without this, if the restservice is up, requests that are run later,
will face a restservice that has already been started for a few
seconds, so those requests will not have to wait as much.
(the next request in line is uploading the license, if provided;
otherwise, I think setting up the CLI, but there might be something
before that).

And if the service is not up, one of those requests will fail instead.
We'll still know about it, but the error will be perhaps less
beautiful (but more useful, if you like getting the underlying
errors).

All in all, taking into account that I don't remember this check
_ever_ failing, I'm going to say that dropping this check for a
15-20% speedup (in the overall container start process) is worth it

* Update CI to build the new image

Having to wait for the 'background' service to finish is
unfortunate. I don't have a better idea.

* Move CLI up in the service ordering

Setting up the CLI earlier means that the manager is usable
earlier. Which, it's not fully usable yet, because the mgmtworker
isn't started yet - so it can't run workflows, but it can queue them.

It's just a bit better usability, to give the user something that
works ASAP.

* In an AIO image, install all services ever.

This allows us to only ship one image, and use it in any role:
AIO, manager, db, or queue. The image will be a bit bigger, because
it includes eg. patroni, but that's a good price to pay, since
those cluster-only packages are negligible in size compared to the
image total

* Set rabbitmq CA_PATH in both configure/start

It is used in start, so it has to be set there.
We don't set it just "always", because apparently it needs to be
NOT set when `_generate_rabbitmq_certs` is called. Amazing.

I'm not going to refactor the whole thing just yet.

* Postgresql_server: stop useless restarts

No need to restart this many times

* Implement `cfy_manager wait-for-starter`

Waiting for the starter service to finish is going to be
a common theme, so it's going to be useful to have a python
implementation, instead of having to conjure up some bash,
which is going to be less useful anyway

* Rewrite container spawns in CI

Spawn the containers as we go, not all up front, and give them
pre-determined IPs.

Instead of using `docker cp` to move files into the containers, and
then running `cfy_manager install`, we just mount the files and
that's it. Magical.

* Ignore manager status in the `cfy cluster status` check

Fixing this is what CY-2823 is about