Add a jenkins pipeline for building the main rpm (#792)
* Add a jenkins pipeline for building the main rpm

We can use this in jenkins, instead of writing the job
configuration there. It's better to have scripts like this in
source control.

This is using a locally-built docker image for building the
rpm, instead of the rpmbuild/centos7 image, because:
1) the rpmbuild/centos7 image insists on running as a specific
  username/uid, and jenkins insists on running things as a different
  uid. So we `-u root` and run as root inside of the container instead
2) that image is not very official, and is not maintained by rh

* install_rpm.spec: add gcc build-requirement

it is actually required for building our rpm.
The rpmbuild/centos7 image just happened to include it.