CY-3235 Print out admin credentials after installation
The current printout is in awkward places. Anyway, there's just
two ways the user is going to do an unconfigured or semi-configured
install, and that's `cfy_manager install` when installing from
a rpm, or `cfy_manager wait-for-starter` when waiting on a container.

And in those cases, tell the user their password.

Other cases, such as `cfy_manager configure`, are useless

Also, if we want to print to screen, just print it, no reason for
the stupid logs/level/handlers dance