Refactor wait_for_starter (#879)
* don't setup_logger at module level

I don't ALWAYS want to log to the logfile

* wait-for-starter: no _prepare_execution

wait-for-starter itself doesn't need to log to the logfile,
it's the thing that READS the logfile. Having it log there just leads
to duplicated error messages.

* follow the log by doing `tail -F`

instead of separate methods for systemd/supervisord, just tail
the log.

Also, previous supervisord log tailing tailed supervisord's own log,
which is less than useful for looking at the manager bootup

* systemd starter: throw on errors

this will make wait-for-starter return nonzero when the starter service
failed

* don't print wait-for-starter's useless tracebacks

tracebacks from subprocesses are very uninteresting and are just
hiding the underlying cause. The actual real traceback, which is
from `cfy_manager start` (or configure) is going to be visible in
the log and in the tail

* refactor wait-for-starter

a common `while deadline` loop, only separate check functions
for systemd/supervisord

* no need for the sys.exit in this case

bootstraperror is already formatted nicely by the excepthook

* systemd: use ExecMainStatus

the other thing is... no idea what and what for. 1 just means 'exited'
there. Systemd doesn't document this very well.

* sleep a second after finishing to allow tail to show the last logs

* formatting

* tune the tail command

max-unchanged-stats to catch truncated/rewritten file

* implement a `tail -F` in python

the gnu one is way unreliable on some filesystems