Default services_to_install to services already installed
If some services are already installed, default the service
list to that. This is so that if you first install some services,
and then run `cfy_manager configure`, then you don't need to
provide services_to_install, because it will be defaulted to
what you already installed.

The list is still overridden by the config.yaml, as before.