Split to cloudify_manager and cloudify_monitoring
Two policies have been defined: one for the Cloudify Manager, the other
one for the Status Reporter (monitoring).  Monitoring policy is defined
for ports 9090-9094 regardless of the fact that only 9090 is actually
used.  This is because SELinux does not allow a single port 9090 to be
redefined, as it is already defined as part of the `websm_port_t`.