RD-673 Refactor the "initialize db" flow
The "is manager in db" check is not enough for assuming that we're
in a cluster. Instead, look at how many managers we have, and only
assume we're in a cluster if there's already a manager other than us.

Also rework the db utils to just return the hostnames, instead of
a separate "is hostname in db" and "get count" utils.