RD-673 Refactor the "initialize db" flow (#1000)
* RD-673 Refactor the "initialize db" flow

The "is manager in db" check is not enough for assuming that we're
in a cluster. Instead, look at how many managers we have, and only
assume we're in a cluster if there's already a manager other than us.

Also rework the db utils to just return the hostnames, instead of
a separate "is hostname in db" and "get count" utils.

* only call insert_manager once

that call was duplicated inside of join_cluster, but of course we
only want to call it once

* move handle_ca_cert up

why was it ever in _initialize_db?! it has nothing to do with the db.

* move the get_monitoring_config bits to a separate script

i want this to be separate from the main
create_tables_and_add_defaults stuff. It's cleaner this way.

Also flatten the _update_config function one level right there.

* drop unused status_reporter_user bits