update_stored_manager: update stored RabbitMQBroker and DBNodes
For AIO, RabbitMQBroker and DBNode point to the local manager,
and need to be updated as well. Simple as that.

Filter by hostname + IP, so that we only find AIO nodes indeed.

Perhaps only the hostname would've been enough - but if the IP
doesn't match, then we can't do anything about it anyway, because
that means the user must've changed it themselves, and they presumably
want to keep it that way.