prometheus.py: a function to get cluster config
Instead of updating the global config, use a function that returns
the config, and use that.

We can't just go and update the config, because specifically,
adding .postgres.nodes will break later use, in case where
there isn't a db cluster (eg. a manager with a single external db)
(eg. rds)