RD-821 monitoring: stop overwriting the config (#1019)
* prometheus.py: a function to get cluster config

Instead of updating the global config, use a function that returns
the config, and use that.

We can't just go and update the config, because specifically,
adding .postgres.nodes will break later use, in case where
there isn't a db cluster (eg. a manager with a single external db)
(eg. rds)

* get_cluster_config: drop unused rabbitmq_ca

i don't see any place where it's used...?

* refactor get_monitoring_config.py

we control both sides of this script, so we might as well
use a format that is nicer to use