get_monitoring_config.py: don't break on the DB being down
If we cannot connect to the DB, just return empty. Then, if we
can't ask the DB what should we be monitoring, we will only
be monitoring the nodes provided in config.yaml.

This only really happens in remove, when we're removing one
service (think 3-node cluster), and then possibly we still want
to monitor some other services, that is why we update
the monitoring targets in remove.

But if we removed the db, we just can't ask it what to monitor.