Fix upgrade in cluster mode (#1031)
* Fix upgrade in cluster mode.

On a 3 nodes cluster, the postgres_exporter.service configuration is wrong. To fix it, the `set_globals()` function was added. 
Both on a 3 nodes cluster and a 9 nodes cluster, CLUSTER_JOIN would be False since the upgrade process doesn't go through the restservice component that configures it. So, an `upgrade` flag was added that specifies if the Prometheus configuration is done as a part of an upgrade or not.