Run components' upgrade functions on upgrade (#1036) (#1038)
* Re-configure Stage component upon upgrade

* Re-configure Composer component upon upgrade

* Re-configure UI before stopping DB

During the upgrade procedure re-run UI components configuration when
database service is still running.  This is because database service is
required for the configure() functions of Stage and Composer components.

* Add extra function for upgrading components

* Prometheus().upgrade() implementation

* Stop the components in revesed order

It might be a good idea to stop the services in reverse order from how
they are setup/started.  Mainly due to their dependencies (on db service
mostly).