RD-1867 CLI: use and update a single profile (#1103)
* cli: use+set

To allow changing the manager IP while staying at the same CLI
profile (to keep the password etc), designate the "manager-local"
name and use that.
Instead of a `cfy prof use`, have separate `cfy prof use` and then
a `cfy prof set`.

* factor out the loglevel mangling

* remove unused method

* remove the start call

start does nothing for the CLI, no need to do it at all. Only thing
it does is emits useless logs.

* cli: only pass username/password when provided

it'll often be the case that the password is not provided. Then,
just don't pass it in, the already-stored password will be used.

* when a cli command fails, show the stdout+stderr

i'm sick of not seeing it

* also add `--ssl on`

yes it's required.