Don't rerender Prometheus configuration on upgrade
In case of an upgrade from previous installation of cloudify-manager,
let's not re-render Prometheus configuration files.

This is because an upgrade to the cloudify-manager software should not
modify the configuration of its components, rather upgrade the software
itself.

In case of a direct requirement for a change, it should be handled by
the RPM file itself.