RD-2146 Using is_not operator in services system filter (#1144)
The is_not operator was created especially for this task. We would like the services' system filter to capture all the deployments that have a label with key csys-obj-type and value other than environment, or don't have a label with the key csys-obj-type at all. This is done using the is_not operator.