Fix container restart (#1166)
* RD-1858 Allow `cfy_manager configure` to change the password (#1102)

* update-stored-manager: also update the admin password

This ONLY changes the password. As opposed to the update-admin-password
command, it doesn unlock the account or reset the login counter or
nothing.

* cli: cope with changing passwords

when the password is changed, the profile must be recreated

* remove now-unused notes and checks

* image-starter: don't default the password if already configured

* RD-1867 CLI: use and update a single profile (#1103)

* cli: use+set

To allow changing the manager IP while staying at the same CLI
profile (to keep the password etc), designate the "manager-local"
name and use that.
Instead of a `cfy prof use`, have separate `cfy prof use` and then
a `cfy prof set`.

* factor out the loglevel mangling

* remove unused method

* remove the start call

start does nothing for the CLI, no need to do it at all. Only thing
it does is emits useless logs.

* cli: only pass username/password when provided

it'll often be the case that the password is not provided. Then,
just don't pass it in, the already-stored password will be used.

* when a cli command fails, show the stdout+stderr

i'm sick of not seeing it

* also add `--ssl on`

yes it's required.

* store metadata for the external cert (#1105)

* store_cert_metadata: rename an argument

it's not necessarily an ip, and it's not necessarily private

* external cert: default to generating from the internal CA

if external CA is not provided, use the internal CA

* store a separate metadata file for the external cert

* re-create the external cert if it doesn't cover the new ips

only if it was auto-generated in the first place (only then will
the metadata file exist). This is similar to _internal_certs_exist

* CLI: use the CA path

now the auto-generated cert is not self-signed, so provide the
CA that we do sign it with

* accidentally a word

* RD-2382 Unify retries behavior on create_db_monitoring_user_query (#1164)

Co-authored-by: Lukasz Maksymczuk <tehasdf@users.noreply.github.com>
Co-authored-by: glukhman <47591609+glukhman@users.noreply.github.com>