jenkins: move creating a cluster to a separate .sh (#1176)
* Jenkinsfile: move starting the cluster to an outside script

This script is going to be useful to be called by hand as well.
And it's easier to maintain outside of the jenkinsfile anyway.

* add the usual bash prelude

* parametrize the script

instead of jenkins-specific ${env} references, use arguments to the
script, and have reasonable defaults as well

* prefix all rendered yaml files with rendered_

this will make it easier to remove them

* remove unused `export`s

create_certs is sourced anyway, not executed

* put everything in functions

i dont really want to refactor it further right now.
But this is a start.

* move everything into a tmpdir

makes for way easier cleanup

* create container names from prefix for validate_status.sh

* jenkinsfile: drop the "cluster image" dockerfile

not needed. We are just using the same image anyway. That one does
`cfy_manager install --only install` in it.

* drop unused file

it even referenced a file in .circleci/ which doesn't exist

* drop .circleci

well unused for quite a while