Modify workers_count calculation
Because we are going to use both gunicorn and uvicorn on the same
machine, the auto-calculation needs to be slightly different.  Instead
of going with `worker_count = number_of_cpus * 2 + 1` equation, let's go
with `worker_count = number_of_cpus * cpu_ratio + 1`.

`cpu_ratio` is a new parameter for `restservice.{gunicorn,uvicorn}`.