RD-4281 Stop duplicating supervisord config (#1326) (#1331)
We already have supervisord.conf! No need to have a docker-specific
one.

The programs that ARE specific, might well live in a separate file.

I note that the docker supervisord.conf used to be SLIGHTLY different
from the rpm one: the logging level is now info, and we do declare
`user=root`. It was root anyway, just not declaring it was a crit-level
warning on supervisord startup.