config: store absolute paths to the config files
TBH the idea of "pass filenames and they're in /etc/cloudify" is
stupid anyway. But I'm not gonna get rid of it entirely, we'll
still require that config files are there. But I'm just gonna
allow passing absolute paths to them as well.

And let's store absolute paths too. It'll make it easier to pass them
to external scripts.