RD-6073 Use configure_manager to create the admin user (#1429)
* RD-6073 Use configure_manager to create the admin user

Instead of doing this in the script here, call out to the module
in the restservice venv.

In the future, we'll move more and more things out of that script,
into the manager module. But for right now, only this.

* config: store absolute paths to the config files

TBH the idea of "pass filenames and they're in /etc/cloudify" is
stupid anyway. But I'm not gonna get rid of it entirely, we'll
still require that config files are there. But I'm just gonna
allow passing absolute paths to them as well.

And let's store absolute paths too. It'll make it easier to pass them
to external scripts.

* run_script_on_manager_venv: handle `-m`

We run that module with `python -m`, so let's make that actually work.
The syntax is a bit silly, but eh.

* Rewrite dict_merge to be our own code

I don't really like the idea of using some random gist. Instead, let's
have our own. It's a bit different because it's DFS with an explicit
stack rather than recursion.

We have the same implementation in cloudify-manager, and it is
unit-tested there.

* pass through the admin password separately

this is the value that we might have generated and it's not going
to be in the config file; pass it separately