RD-5986 Drop systemd service management (#1435)
* remove SystemD() backend implementation

Maybe this will want more refactors later to shorten it even more,
but for right now, just removing the whole thing.

* Rewrite haveged&syslog systemd calls

For those cases, we still want to ask systemd, if those services are
system-provided. So just inline these calls essentially.

* Remove all kinds of `if service_management==supervisord`

Now that this is the only way, we don't need the if

* remove service_management from config.yamls

* remove service_management from manager-config

* default syncthing to supervisord

I'm not removing the argument yet, for build-related reasons. I want
this to be able to run as-is, without cloudify-premium or
cloudify-manager changes. I'll remove it later.

* run_init: only run supervisord

Not removing the function just yet, because that's the entrypoint
for existing docker images. I'll remove it ASAP, but not in this PR.
It'll stay here for backwards-compat reasons, but only for a day.

* docker images: make supervisord the entrypoint

* remove systemd-only files

* factor out checking for systemd services

No need for the if-active-but-not-enabled check, either

* syslog restart: use subprocess, not the nonexistent class

whoops

* cache the result of the systemd check

* also do the same for the rsyslog component