RD-6566 Move inserting manager to configure_manager.py (#1453)
* RD-6566 Move inserting manager to configure_manager.py

This is not trivial, and I also refactor a bit, so it's not gonna
be very easy to review; sorry about that, but I wasn't able to come
up with a better way :)

The point is taking that `_insert_manager` out of
create_tables_and_add_defaults.py, and moving it to the
cloudify-manager side.

Alas, that script was also used for inserting a new manager into
an already-existing db, in a cluster. cloudify-manager's
configure_manager.py can do that as well, and does, but for now
I'd like to limit the changes here to a minimum, so I'll
still keep a `def insert_manager` here.

But let's use the `update_stored_manager.py` script for inserting
it. And so, that `update` now really means `upsert`. So I'm refactoring
that script a bit as well.

And restservice.py's `_configure_db` is a bit nicer now as well...

* don't flush manager just yet

don't post it to the db before we attach a cert to it

* allow the "no old cert" case

this happens when just first inserting the manager

* fix tenant query

heh.

* the whole tenant stuff is just unnecessary

tenants are already created in the other script, configure_manager.py!
Doing it here isn't needed.

* no need to set admin password either