Add retries on DB ID check
pg_controldata got unhappy while testing this, but was fine when I then checked it.
This shouldn't be dependent on the amount of data in the DB (because the pg_upgrade
tells you when it has finished upgrading the DB data), but it's not clear what it
is dependent on.